Implement transferable training (#95)
* add jax-dataclasses as a dependency

* include coords in batch, remove them from Molecule

* introduce config_idx to PhysicalConfiguration

For now it's always zero, since there is only one configuration we
sample from. In a transferable setting this field will track which
configuration the given sample is taken from.

* implement transferable training

* [doc] fix train docstring

* fix nanmean and nanstd denominators

* remove unnecessary jnp.asarray from app.py

* fix selection of baseline orbitals during pretraining

* use separate rng keys for sampler initialization

* extend Baseline to multiple configurations

This simplifies the pretraining logic significantly.

* move coords to mol.data

* fix energy logging to result.h5

* add step as config_idx_factory argument

* avoid logging NaNs or infs

* log only the mean of the Hamiltonian stats

* limit sample size of samplers to max_per_config

Remove `smpl_state` argument of `select_idxs`.

* use a single sampler instance for all configurations

* use Molecule objects to store nuclear geometries

* compute mean of local energy molecule-wise

+++ src/deepqmc/molecule.py
@@ -90,7 +90,7 @@ class Molecule:
     pp_nl_params: jnp.ndarray
 
     # REDUNDANT AUXILIARY PROPERTIES:
-    # True if at leas one nucleus uses pseudopotential
+    # True if at least one nucleus uses pseudopotential
     any_pp: bool
     # Indices of nuclei with pseudopotential
     nuc_with_nl_pot: jnp.ndarray
@@ -121,7 +121,7 @@ class Molecule:
         unit_multiplier = {'bohr': 1.0, 'angstrom': angstrom}[unit]
         set_attr(
             coords=unit_multiplier * jnp.asarray(coords),
-            charges=1.0 * jnp.asarray(charges, dtype=jnp.int8),
+            charges=1.0 * jnp.asarray(charges),
             charge=charge,
             spin=spin,
             pp_mask=jnp.array(pp_mask),
